definitivo ya corregido
Con tester...

+++ ft_strlen.c
@@ -6,15 +6,13 @@
 /*   By: ancuesta <ancuesta@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/09 13:28:52 by ancuesta          #+#    #+#             */
-/*   Updated: 2024/06/09 17:58:16 by ancuesta         ###   ########.fr       */
+/*   Updated: 2024/06/13 21:09:24 by ancuesta         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdio.h>
-#include <ctype.h>
-#include <string.h>
+#include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	int	index;
